requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
PyPDF2==3.0.1
beautifulsoup4==4.12.3
requests==2.32.3
anthropic==0.39.0
